Define custom mapping via attribute macro
