+++ examples/dribbble.rb
@@ -1,12 +1,12 @@
 require 'jeckle'
 
-#Here is an example using the Dribbble API and 
+# Here is an example using the Dribbble API
 
 Jeckle.configure do |config|
   config.register :dribbble do |api|
     api.base_uri = 'http://api.dribbble.com'
     api.middlewares do
-    	response :json
+      response :json
     end
   end
 end
@@ -19,7 +19,7 @@ class Shot
   attribute :id, Integer
   attribute :name, String
   attribute :url, String
-  attribute :image_url, String
+  attribute :image_url, String, as: :image
 end
 
 shot = Shot.find 1600459
@@ -30,6 +30,5 @@ puts "The shot id is #{shot.id}"
 puts "The shot name is #{shot.name}"
 # => Daryl Heckle And Jeckle Oates
 
-puts "The shot image URL is #{shot.image_url}"
-
+puts "The shot image URL is #{shot.image}"
 # => https://d13yacurqjgara.cloudfront.net/users/85699/screenshots/1600459/daryl_heckle_and_jeckle_oates-dribble.jpg
