+++ lib/jeckle/errors.rb
@@ -26,30 +26,4 @@ module Jeckle
       super message
     end
   end
-
-  class InvalidAttributeMappingError < ::StandardError
-    def initialize(data)
-      class_name = data[:class_name]
-      resource_attribute = data[:resource_attribute]
-      api_attribute = data[:api_attribute]
-
-      message = %{Invalid attribute mapping: #{resource_attribute} => #{api_attribute}.
-
-        Heckle: - Hey chum, what we can do now?
-        Jeckle: - Old chap, you need to declare the #{resource_attribute} attribute.
-        Heckle: - Hey pal, then tell me how!
-        Jeckle: - Deal the trays, old thing:
-
-        class #{class_name}
-          attribute :#{resource_attribute}
-
-          mapping do
-            attribute :#{resource_attribute}, :#{api_attribute}
-          end
-        end
-      }
-
-      super message
-    end
-  end
 end
