+++ lib/jeckle/resource.rb
@@ -7,7 +7,7 @@ module Jeckle
       base.send :include, Jeckle::HTTP
       base.send :include, Jeckle::RESTActions
 
-      base.send :include, Jeckle::CustomAttributeMapping
+      base.send :extend, Jeckle::CustomAttributeMapping
     end
   end
 end
