+++ spec/lib/jeckle/api_spec.rb
