+++ spec/lib/jeckle/http_spec.rb
