+++ spec/lib/jeckle/model_spec.rb
