+++ spec/lib/jeckle/request_spec.rb
