+++ spec/lib/jeckle/rest_actions_spec.rb
