+++ spec/lib/jeckle/setup_spec.rb
